% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{flip_values_and_names}
\alias{flip_values_and_names}
\title{Flip values and names}
\usage{
flip_values_and_names(x)
}
\arguments{
\item{x}{}
}
\value{
Same structure as before, but names as values and values as names
}
\description{
Flip values and names
}
