% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{is_answer_true_false}
\alias{is_answer_true_false}
\title{Is answer really TRUE or FALSE?}
\usage{
is_answer_true_false(answer)
}
\arguments{
\item{answer}{(\link{character}) Answer}
}
\value{

}
\description{
Is answer really TRUE or FALSE?
}
