% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{is_valid}
\alias{is_valid}
\title{Is valid}
\usage{
is_valid(x, valid_fn, ...)
}
\arguments{
\item{x}{[\link{chacter} or \link{integer}] A valid choice}

\item{valid_fn}{[\link{function}] The actual validation function}

\item{...}{Additional arguments passed to \code{valid_fn}}
}
\value{

}
\description{
Convenience wrapper around using a \verb{valid_*} function directly
}
\examples{
is_valid("yes", valid_yes_no)
is_valid(c("yes", "no"), valid_yes_no)
try(is_valid("YES", valid_yes_no))
is_valid(c("yes", "NO"), valid_yes_no)
}
