% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{not_in}
\alias{not_in}
\title{Check if not in set}
\usage{
not_in(x, set)
}
\arguments{
\item{x}{(\link{character})}

\item{set}{(\link{character})}
}
\value{
(\link{logical(1)})
}
\description{
Check if not in set
}
