% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-package.R
\docType{package}
\name{valid-package}
\alias{valid}
\alias{valid-package}
\title{valid: Facilitating systematic input validation}
\description{
Little helper functions for systematic input validation
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
