% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{valid}
\usage{
valid(x, class1 = class(x))
}
\arguments{
\item{x}{an object}

\item{class1}{a 'character'}
}
\description{
Whereas 'is()' returns TRUE if an
object already has a class, 'valid()' returns TRUE if
an object passes all the validity tests required by
the class. When a valid() method provided, if this 
function returns TRUE, then the class constructor
should not throw an error if called with the same
data, since the constructor uses this code for 
validation.
}
