% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_again_exit}
\alias{valid_again_exit}
\title{Valid: again/exit}
\usage{
valid_again_exit(choice = character(), ...)
}
\arguments{
\item{choice}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}})]}
}
\value{

}
\description{
Valid: again/exit
}
\examples{
valid_again_exit()
valid_again_exit("again")
try(valid_again_exit("invalid"))
valid_again_exit(reverse = TRUE)
valid_again_exit(flip = TRUE)
valid_again_exit(unname = TRUE)
}
