% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_authentication}
\alias{valid_authentication}
\title{Valid authentication}
\usage{
valid_authentication(auth = character(), flip = FALSE)
}
\arguments{
\item{auth}{(\link{character}) Authentication choice}

\item{flip}{(\link{logical}) Flip the vector's order?}
}
\value{

}
\description{
Valid authentication
}
