% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_dep_types}
\alias{valid_dep_types}
\title{Valid dependency types}
\usage{
valid_dep_types(type = character(), ...)
}
\arguments{
\item{type}{(\link{character}) Dependency type choice}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid dependency types
}
\examples{
valid_dep_types()
valid_dep_types("Suggests")
try(valid_dep_types("invalid"))
valid_dep_types(reverse = TRUE)
valid_dep_types(flip = TRUE)
valid_dep_types(unname = TRUE)
}
