% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_devops_envs}
\alias{valid_devops_envs}
\title{Valid DevOps environments}
\usage{
valid_devops_envs(devops_env = character(), ...)
}
\arguments{
\item{devops_env}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid DevOps environments
}
\examples{
valid_devops_envs()
valid_devops_envs("staging")
try(valid_devops_envs("invalid"))
valid_devops_envs(reverse = TRUE)
valid_devops_envs(flip = TRUE)
valid_devops_envs(unname = TRUE)
}
