% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_keep_reset}
\alias{valid_keep_reset}
\title{Valid: keep/reset}
\usage{
valid_keep_reset(choice = character(), ...)
}
\arguments{
\item{choice}{(\link{character}) Actual choice out of all available choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid: keep/reset
}
\examples{
valid_keep_reset()
valid_keep_reset("keep")
try(valid_keep_reset("invalid"))
valid_keep_reset(reverse = TRUE)
valid_keep_reset(flip = TRUE)
valid_keep_reset(unname = TRUE)
}
