% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_licenses}
\alias{valid_licenses}
\title{Valid licenses}
\usage{
valid_licenses(license = character(), ...)
}
\arguments{
\item{license}{\link{character} License choice}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}#'}
}
\value{

}
\description{
Valid licenses
}
