% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_remote_storage}
\alias{valid_remote_storage}
\title{Valid remote storage}
\usage{
valid_remote_storage(storage = character(), ...)
}
\arguments{
\item{storage}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid2]{valid2()}}}
}
\value{

}
\description{
Valid remote storage
}
\examples{
valid_remote_storage()
valid_authentication("ssh")
try(valid_authentication("invalid"))
valid_authentication(reverse = TRUE)
valid_authentication(flip = TRUE)
valid_authentication(unname = TRUE)
}
