% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_storage}
\alias{valid_storage}
\title{Valid storage}
\usage{
valid_storage(storage = character(), ...)
}
\arguments{
\item{storage}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid2]{valid2()}}}
}
\value{

}
\description{
Valid storage
}
