% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_storage_local}
\alias{valid_storage_local}
\title{Valid local storage}
\usage{
valid_storage_local(storage = character(), ...)
}
\arguments{
\item{storage}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid2]{valid2()}}}
}
\value{

}
\description{
Valid local storage
}
\examples{
valid_storage_local()
valid_storage_local("fs")
try(valid_storage_local("invalid"))
valid_storage_local(reverse = TRUE)
valid_storage_localn(flip = TRUE)
valid_storage_local(unname = TRUE)
}
