% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_storage_remote}
\alias{valid_storage_remote}
\title{Valid remote storage}
\usage{
valid_storage_remote(storage = character(), ...)
}
\arguments{
\item{storage}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid2]{valid2()}}}
}
\value{

}
\description{
Valid remote storage
}
\examples{
valid_storage_remote()
valid_storage_remote("gcp_s3")
try(valid_storage_remote("invalid"))
valid_storage_remote(reverse = TRUE)
valid_storage_remoten(flip = TRUE)
valid_storage_remote(unname = TRUE)
}
