% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_yes_no}
\alias{valid_yes_no}
\title{Valid: yes/no}
\usage{
valid_yes_no(choice = character(), ...)
}
\arguments{
\item{choice}{(\link{character}) Actual choice out of all available choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid: yes/no
}
\examples{
valid_yes_no()
valid_yes_no("yes")
try(valid_yes_no("invalid"))
valid_yes_no(reverse = TRUE)
valid_yes_no(flip = TRUE)
valid_yes_no(unname = TRUE)
}
