% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_yes_no_again_exit}
\alias{valid_yes_no_again_exit}
\title{Valid: yes/no/again/exit}
\usage{
valid_yes_no_again_exit(choice = character(), ...)
}
\arguments{
\item{choice}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid: yes/no/again/exit
}
\examples{
valid_yes_no_again_exit()
valid_yes_no_again_exit("yes")
try(valid_yes_no_again_exit("invalid"))
valid_yes_no_again_exit(reverse = TRUE)
valid_yes_no_again_exit(flip = TRUE)
valid_yes_no_again_exit(unname = TRUE)
}
