% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{Generic validation function}
\usage{
valid(
  choice = character(),
  choices = character(),
  reverse = FALSE,
  flip = FALSE,
  strict = TRUE
)
}
\arguments{
\item{strict}{}
}
\value{

}
\description{
Generic validation function
}
