% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_authentication}
\alias{valid_authentication}
\title{Valid authentication}
\usage{
valid_authentication(auth = character(), ...)
}
\arguments{
\item{auth}{(\link{character}) Selection from available valid choices}

\item{...}{Further arguments that will be passed to \code{\link[=valid]{valid()}}}
}
\value{

}
\description{
Valid authentication
}
\examples{
valid_authentication()
valid_authentication("ssh")
try(valid_authentication("invalid"))
valid_authentication(reverse = TRUE)
valid_authentication(flip = TRUE)
valid_authentication(unname = TRUE)
}
