% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_dep_types}
\alias{valid_dep_types}
\title{Valid: dependency types}
\usage{
valid_dep_types(type = character(), flip = FALSE)
}
\arguments{
\item{type}{(\link{character}) Dependency type choice}

\item{flip}{(\link{logical}) Flip the vector's order?}
}
\value{

}
\description{
Valid: dependency types
}
