% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_devops_envs}
\alias{valid_devops_envs}
\title{Valid DevOps environments}
\usage{
valid_devops_envs(devops_env = character(), reverse = FALSE, flip = FALSE)
}
\arguments{
\item{devops_env}{(\link{character}) DevOps environment choice}

\item{reverse}{(\link{logical}) Reverse the vector's order?}

\item{flip}{(\link{logical}) Flip names and values?}
}
\value{

}
\description{
Valid DevOps environments
}
\examples{
valid_devops_envs()
valid_devops_envs("staging")
valid_devops_envs("staging", reverse = TRUE)
}
