% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_licenses}
\alias{valid_licenses}
\title{Valid licenses}
\usage{
valid_licenses(license = character(), flip = FALSE)
}
\arguments{
\item{license}{\link{character} License choice}

\item{flip}{(\link{logical}) Flip the vector's order?}
}
\value{

}
\description{
Valid licenses
}
