% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_none}
\alias{valid_none}
\title{Valid: none}
\usage{
valid_none(choice = character(), flip = FALSE)
}
\arguments{
\item{choice}{\link{character} Actual choice out of all available choices}

\item{flip}{(\link{logical}) Flip the vector's order?}
}
\value{

}
\description{
Valid: none
}
