% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_yes_no_again_exit}
\alias{valid_yes_no_again_exit}
\title{Valid: yes/no/again/exit}
\usage{
valid_yes_no_again_exit(choice = character(), flip = FALSE)
}
\arguments{
\item{choice}{\link{character} Actual choice out of all available choices}

\item{flip}{(\link{logical}) Flip the vector's order?}
}
\value{

}
\description{
Valid: yes/no/again/exit
}
